requirements.txt
absl-py==1.4.0
altgraph==0.17.3
astor==0.8.1
cachetools==4.2.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
colorama==0.4.6
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
h5py==2.10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
ISR==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
ordered-set==4.1.0
pefile==2023.2.7
Pillow==9.5.0
protobuf==3.20.1
pyaml==23.5.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
PyMuPDF==1.22.5
pywin32-ctypes==0.2.2
PyYAML==6.0.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==2.3.0
tqdm==4.66.1
ttkbootstrap==1.10.1
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zstandard==0.21.0
